Add All Tabs with Content
This commit adds all tabs and their respective content to the webpage using webpack. The tabs all connect to one tabs.js and tabs,js connects to a pageload file, which then connects to the index file. The content is all made through javascript, with the excpetion of the header and footer, which do not change. The itineraries are being made by a function for each day, compiling all of them into one function, and then that function is being used in another function which creates the rest of the DOM.